jQuery Scala Stack - Routing refactoring - Renaming - New packages
