+++ src/main/webapp/scalajs/shared/Route.scala
@@ -1,12 +1,27 @@
 package shared
 
-object Route {
+import shared.html._
+import org.slf4j.LoggerFactory
+import org.slf4j.Logger
+import scalatags.Text.TypedTag
+import org.glassfish.jersey.uri.UriTemplate
 
-  def apply(route: Option[String] = None) = {
-    route match {
-      case Some(route) if route == "/" | route.startsWith("/index") =>
-        Html(new Stylisch, route.endsWith(".min.html"))
-      case _ => Four04()
+object Route {
+  
+  def apply(uri: String)(implicit logger : Logger = LoggerFactory.getLogger(getClass)) : Either[String, TypedTag[String]] = {
+    
+    val map = new java.util.HashMap[String, String]();
+    val template = new UriTemplate("/{language}/{page}");
+    template.`match`(uri, map)
+    
+    logger.info(s"mapping $map")
+    
+    uri match {
+      case "/" => Left(s"/${Languages.default.code}/")
+      case route if route.startsWith("""/\s\s-\s\s/""") => Right(Index(new Stylisch, minified = route.endsWith(".min.html")))
+      case "/404" => Right(ClientError.NotFound())
+      case "/5xx" => Right(ServerError.InternalServerError())
+      case _ => Left("/404")
     }
   }
 }
